Update tools.md
Noticed a typo in the file, so I fixed it.  I also cleaned up the layout of the file for consistency sake (I found a couple of extra line breaks). Replaced JS with JavaScript everywhere (because newbies are not going to know what an arbitrary abbreviation is - besides, why abbreviate JavaScript when you have the space to spell it out completely).